refactor : 특정 욕설 및 선정적인 단어 사용 불가하게끔 변경 및 env로 이동 완료 #67
