+++ app/(root)/auth/kakao/page.tsx
@@ -1,37 +1,17 @@
 'use client'
 
-import { redirect, useSearchParams } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect } from 'react';
+import { getKakaoAuthCode } from '../../../../api/auth/getKakaoAuthCode';
 
 export default function KakaoAuth() {
   const searchParams = useSearchParams();
+  const router = useRouter()
   const code = searchParams.get('code');
 
   useEffect(() => {
     if (code) {
-      fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/kakao?code=${code}`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: 'include',
-      })
-      .then(response => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! Status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then(data => {
-        console.log('Success:', data);
-        localStorage.setItem('signupData', JSON.stringify(data));
-      })
-      .then(
-        redirect('/sign-up')
-      )
-      .catch((error) => {
-        console.error('Error:', error);
-      });
+      getKakaoAuthCode(code, router)
     }
-  }, [code]);
+  }, [code, router]);
 }
