+++ components/boxes/PostBox.tsx
@@ -10,7 +10,7 @@ import { useSelector } from 'react-redux';
 import { RootState } from '../../store/store';
 import { selectNumberOfBoxes } from '../../store/slice/boxLayoutSlice';
 
-export default function  PostBox({ item }: PostBoxNewProps) {
+export default function  PostBox({postId, photoId, photo_url, like, saved }: PostBoxNewProps) {
 
   const numberOfBoxes = useSelector((state: RootState) => selectNumberOfBoxes(state));
 
@@ -26,10 +26,11 @@ export default function  PostBox({ item }: PostBoxNewProps) {
 
   return (
     <div className="box relative mb-[16px] flex items-center justify-center rounded-2xl overflow-hidden transition-all duration-500  border-[5px] border-transparent hover:border-purple p-[5px] snap-start" style={{ width: postBoxWidths[layoutNum]}}>
-      <Link href={`post/${item.data.postId}`} className="absolute inset-0 z-0">
+      <Link href={`post/${postId}`} className="absolute inset-0 z-0">
         <Image
-          src={'https://helpx.adobe.com/content/dam/help/en/photoshop/using/convert-color-image-black-white/jcr_content/main-pars/before_and_after/image-before/Landscape-Color.jpg'}
-          alt="postImage"
+          src={`https://helpx.adobe.com/content/dam/help/en/photoshop/using/convert-color-image-black-white/jcr_content/main-pars/before_and_after/image-before/Landscape-Color.jpg`}
+          alt={`postImage`}
+          id={`${photoId}`}
           fill
           style={{objectFit: 'cover', width: '100%', height: '100%'}}
           quality={100}
@@ -40,12 +41,12 @@ export default function  PostBox({ item }: PostBoxNewProps) {
       <button
         onClick={handleHeartClick}
         className="absolute top-2 right-2 z-10">
-        <HeartButton width="21px" height="19px" isClicked={ item.data.like } />
+        <HeartButton width="21px" height="19px" isClicked={ like } />
       </button>
       <button 
         onClick={handlePlusClick}
         className="absolute bottom-2 right-2 z-10">
-        <PlusButton width="24px" height="24px" isClicked={ false }/>
+        <PlusButton width="24px" height="24px" isClicked={ saved }/>
       </button>
     </div>
   );
