+++ types/dataType.ts
@@ -11,11 +11,23 @@ export type UserData = {
 };
 
 export type SignUpData = {
-    success: boolean,
-    message: string,
-    kakaoUserInfoDto: {
-        id: number,
-        email: string,
-        platform: string
-    }
+  success: boolean,
+  message: string,
+  kakaoUserInfoDto: {
+      id: number,
+      email: string,
+      platform: string
+  }
 }
+
+export type PostsData = {
+  success: boolean;
+  code: number;
+  data: Array<{
+    postId: number;
+    photoId: number;
+    photo_url: string;
+    like: boolean;
+    saved: boolean;
+  }>;
+};
